Move ExtractError to lib
